making everything work with the new modules
