+++ irods_testing_environment/irods_config.py
@@ -6,9 +6,9 @@ import logging
 import os
 
 # local modules
-import context
-import execute
-import json_utils
+from . import context
+from . import execute
+from . import json_utils
 
 def get_irods_zone_name(container):
     """Return the Zone name of the iRODS server running on `container`."""
